battery_agent/requirements.txt
anthropic>=0.40.0
pandas>=2.0.0
streamlit>=1.32.0
reportlab>=4.0.0
python-dotenv>=1.0.0
numpy>=1.26.0
rich>=13.0.0
